mvcc: abort duplicates after rollback This commit fixes a Memtx-MVCC related bug that could lead to duplicates in secondary indexes after rollback.
To guarantee the absence of duplicates in secondary indexes, MVCC
maintains the following invariant for all in-progress transactions:
```
If an in-progress story `x` conflicts with a story `y` in some secondary
index (i.e., they have the same key in that index), then `x` must also
conflict with `y` in the primary key.
```
`x` and `y` may belong to the same transaction or different ones. The case
where `x` and `y` belong to the same transaction is trivial. If `x` and `y`
belong to different transactions, then `y` must be the last prepared story
in the chain corresponding to that index. This implies that the invariant
may break for some transactions when the last prepared story in a chain
changes (either when another story becomes last or when the last story's
`del_psn` becomes 0). All such cases must be handled - any transactions
that violate the invariant (i.e., start duplicating prepared tuples) must
be aborted.

Rollback often leads to changes in the last prepared story within chains.
However, this case was previously overlooked, which could result in
duplicates after rollback. This commit adds the missing handling for
rollback scenarios.